// src/type/global.type.ts
export interface Movies {
    Title: string;
    Year: string;
    imdbID: string;
    Type: string;
    Poster: string;
}

type Rating = {
    Source: string;
    Value: string;
}

export interface Movie {
    // [key: string]: string | Rating[];
    Title: string;
    Year: string;
    Rated: string;
    Released: string;
    Runtime: string;
    Genre: string;
    Director: string;
    Writer: string;
    Actors: string;
    Plot: string;
    Language: string;
    Country: string;
    Awards: string;
    Poster: string;
    Ratings: Rating[];
    Metascore: string;
    imdbRating: string;
    imdbVotes: string;
    imdbID: string;
    Type: string;
    DVD: string;
    BoxOffice: string;
    Production: string;
    Website: string;
    Response: string;
    totalSeasons: string;
}

export interface MovieData {
    movieData: Movie;
}